Cleaned code, exceptions updated
