+++ pyproject.toml
@@ -75,8 +75,8 @@ exclude = [
 line-length = 96
 indent-width = 4
 [tool.ruff.lint]
-select = ["E", "F", "B", "I", "N", "UP", "YTT", "ANN", "S", "A"]
-ignore = ["ANN102"]
+select = ["E", "F", "B", "I", "N", "UP", "YTT", "ANN", "S", "A", "D"]
+ignore = ["ANN102", "D100", "D101"]
 
 [tool.ruff.format]
 quote-style = "double"
