--- src/infrastructure/exceptions/infrastucture_base_exceptions.py
@@ -1,2 +1,0 @@
-class InfrastructureBaseError(Exception):
-    """Base class for all exceptions in the infrastructure layer."""
