+++ src/infrastructure/jm_integration/jm_settings.py
@@ -6,9 +6,16 @@ BASE_DIR = Path(__file__).resolve().parent.parent.parent.parent
 
 
 class JMSettings(BaseSettings):
+    """Settings for the JM application.
+
+    These settings are loaded from the `.jm_env` environment file.
+    """
+
     JM_TOKEN: str
 
     class Config:
+        """Configuration for the JM settings."""
+
         env_file = ".jm_env"
         env_file_encoding = "utf-8"
